+++ lib/watir/web_scraper/step.rb
@@ -4,9 +4,10 @@ module Watir
   module WebScraper
     # Represents an atomic step of navigating inside a browser page
     class Step
-      def initialize(browser, params = {})
+      def initialize(browser, params, fetched_data)
         @browser = browser
         @params = params
+        @fetched_data = fetched_data
       end
 
       def call
@@ -25,7 +26,7 @@ module Watir
         raise NotImplementedError, 'This is an abstract base method. Implement in your subclass'
       end
 
-      attr_reader :browser, :params
+      attr_reader :browser, :params, :fetched_data
     end
   end
 end