+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change log
 
-## Unreleased
+## 0.3.0 (2020-03-20)
 
 ### Changes